Schema: change active_storage_variant_records.blob_id to bigint
References #304

+++ db/schema.rb
@@ -43,7 +43,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_12_044725) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.integer "blob_id", null: false
+    t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
